feat: 修复启动bug, 重构question.py
